Refactors pkg/testutils
Solves #11579.
